fix(bench): branchy seeded its data with signed overflow, so its own oracle disagreed
`runtime-bench-check` builds each kernel twice -- once with the reference
compiler to produce the expected output, once with mcc -- and compares. For
branchy the two never agreed:

    want: branchy 487419720 122294685.000000
    got:  branchy -7621192680 -921869400.000000

Neither compiler is wrong. `data[i] = ((i * 1103515245 + 12345) >> 5) % 2001
- 1000` overflows `int`, clang exploits that from -O1, and mcc wraps at every
level; mcc matches `clang -O0` and `clang -fwrapv` to the digit. The kernel
was disagreeing with its oracle about the input data before a single branch
was measured.

Computing the LCG in `unsigned` and converting after makes the seed
well-defined without changing the distribution a wrapping implementation
produces. mcc -O2, clang -O0, clang -O2 and gcc -O2 now all print
`branchy 20309155 5095611.875000`.

This is the fifth of the seven standing reds to close today. It is unrelated
to the absent `vendor/plb` that the audit residue blames for the
runtime-bench family; that one is still open and still makes
`runtime-bench-gatewin` a permanent skip.

+++ tests/runtime/branchy.c
@@ -12,7 +12,7 @@ int main(int argc, char **argv)
 	long acc = 0;
 	double facc = 0.0;
 	for (i = 0; i < (1 << 16); i++) {
-		data[i] = ((i * 1103515245 + 12345) >> 5) % 2001 - 1000;
+		data[i] = (int)(((unsigned)i * 1103515245u + 12345u) >> 5) % 2001 - 1000;
 		fdata[i] = data[i] * 0.125;
 	}
 	for (r = 0; r < n; r++) {
